IMPALA-14095: Calcite Planner: Use pruned row counts in stats
This commit uses the more accurate pruned row counts when collecting
filter cardinalities.

The pruned row counts are cached in the CalciteTable.

The ImpalaRelMdRowCount checks to see if the underlying RelNode
has a Table attached to it.  The condition is broken up into
portions used by the HdfsPartitionPruner for pruning and those
conditions which are not used for pruning.  The conditions not
used in the pruning still need to be calculated in the filter
selectivity estimates.

Change-Id: I442667a338d30d3daa66e223154bc9a1835b2880