IMPALA-14434: Calcite planner: implement partition key scan optimization
Implemented the partition key scan optimization for Calcite planner

Most of the code was already in place. Just needed to refactor some
code in SingleNodePlanner to make it callable from Calcite and use
the already created isPartitionKeyScan method in ImpalaHdfsScanRel.

Testing was done by running the Impala e2e tests with the
use_calcite_planner flag set to true.

Change-Id: I7b5b8a8115f65f6be27a5be0e19f21eebab61a32