IMPALA-14436: Calcite planner: Fix min/max runtime filters
This patch fixes an issue where min/max runtime filters were not getting
created for nested loop joins. The conjuncts need to be placed in
the right variables when initializing the NestedLoopJoinNode.

For inner joins, all the conjuncts need to be in the filterConjuncts
in order to be processed for the min/max runtime filter.

This patch is needed to allow overlap_min_max_filters.test to succeed
for the Calcite planner.

An extra minor fix has been thrown into the patch: When the conjunct
is "IS NOT DISCTINCT FROM", this should be treated as an equi-join
conjunct.

Change-Id: I4e5f9be67a775a771c0868ac4407410385e8a755