IMPALA-14436: Calcite planner: Implement single row join optimization
The original Impala planner has some optimizations for when there is
a single row produced by the right side.

This commit adds code to detect cases where that happens via either an
aggregate that produces one row or a values clause.

If detected, the Join Node is produced in accordance to what is expected
by the original planner.

Change-Id: I4e5f9be67a775a771c0868ac4407410385e8a755