IMPALA-14482: Calcite planner: order by "random()" fixed
The SortRemoveConstantKeysRule in Calcite was removing functions
that did not have any column expressions in it.

This commit makes a modification to that rule to ensure that the
order by is not optimized out for non-deterministic functions.

The test for this can be found in:
test_sort.py::TestRandomSort::test_order_by_random

Change-Id: I6ebf3aa11030d98b447d08d1364f5f65eb2ca661