IMPALA-14485: Change test files to be compatible with Calcite results
Some of the results for the Calcite planner are similar to the original
Impala planner but the test files need to be modified so the tests
pass for both planners.

In one case, a STRAIGHT_JOIN was added to the join to keep the join ordering
consistent.

Some regex was handed because the column name produced by
Calcite is slightly different from the name produced by Impala.

Change-Id: I27864ed9ac0ca51ae4b493f2e02d995a5f521732