IMPALA-XXXXX: Maybe combine with v1.40 upgrade
The v1.40 upgrade adds projects with unions. Let's test this with
and without the v1.40 upgrade because things might break without
this fix. But it might just be that this only fixes
test_const_scalar_expr_in_union_table, which contains a profile
which squishes constants into one union Impala PlanNode.

Change-Id: Ib3a014329a5a9359051809f0d82a518291ee2ee8