IMPALA-14488: Calcite planner: Support fallback to Original planner
If a query fails in Calcite, the query sometimes needs to fallback to
running with the original planner. There are 3 different ways this
commit allows a query to fallback with the query option CALCITE_FALLBACK:

1) nonquery exceptions only.  Calcite currently only supports queries
(no DDL, DML, or other non-select statements).  Running with this option
will throw an exception for every query that Calcite does not support,
including explicit UnsupportedException exceptions.

2) Unsupported and nonquery exceptions:  Similar to 1), but if there is
a failure that we know explicitly fails in Calcite, an UnsupportedException
is caught and we fallback to the original planner. This option is useful on
data load because of some of the data loading uses COMPUTE STATS and under
the covers, it runs stats on complex query columns which we do not want to
fail.

3) All exceptions:  Any exception from the Calcite planner will result in
a retry with the original planner.  This is useful in production scenarios
where we don't want a Calcite bug to be the cause of a failed query.  Normal
jenkins test runs that run with the Calcite planner will most likely use
this option.

Change-Id: I6413b478fd11b645d05a60313990cb2b8c015a67