IMPALA-14222: Calcite planner: Fix exception for unused group in grouping set
An exception was being thrown when "group by grouping sets" was
being used and there was an unrepresented group in the grouping set.

Calcite creates a LogicalUnion for the duplicate grouping set and the
aggregate under this LogicalUnion only contains the duplicate grouping
set, so all the groups are not represented.

The fix was simply to allow null groups when building up the aggregate.
The Impala structures are already fine with this so the change only
needed to be in the Impala Calcite code.

Having said that, queries with duplicate grouping sets do not return
the proper results in the original planner. IMPALA-14483 has been filed
for this.

Change-Id: Ie133733f804d7ed6a461ac8c0bf25cac996635de