IMPALA-14167: Calcite planner: Handle literals and time types in same function
Fixed an issue with finding the best type when a string literal is in the
same function as a date or time literal.

A string literal is treated as a CHAR in Calcite and a STRING in Impala. This
conversion is ultimately made when the abstract syntax tree (SqlNodes) have
been converted into a logical tree (RelNodes).

At validation time, however, the string literal is still treated as a CHAR type.
To further complicate the issue, Impala does not treat the CHAR type and
the date and time types as compatible. However, it does allow compatibility
between CHAR to string and string to date and time.  This issue will not
be dealt with at this stage as it has a huge impact.

For the Calcite case, since we know a string literal is really a string, we
treat CHAR and date and time types as compatible and use the "wider type"
of the date or time type as the compatible type.

Change-Id: I9482b5d1eb42b37533b77d87598f9230b38beba4