IMPALA-XXXXX: Calcite Planner: Fix failure with no equality join condition
The following query fails after upgrading to 1.41:

SELECT a.string_col ts2
FROM alltypes a
     FULL OUTER JOIN (
         select * from alltypessmall where id < 0
     ) b
     ON a.id=b.id AND a.id = 1
ORDER BY ts2 DESC LIMIT 10;

The reason is because a join condition contains =($0,$1), and both are on the
same side. It cannot be pushed because JOIN_CONDITION_PUSH does not allow
the push to the left side. So we need to skip over putting this into the
equality constraints.

Change-Id: Ie826a412bc5e650eb387cd4aa5aacf1a26535931