IMPALA-XXXXX: Calcite Planner: Fix failure with no equality join condition
The following query fails after upgrading to 1.41:

select a.id, b.id, c.id, a.day, b.day, c.month, a.month, b.month
from alltypesagg a join alltypesagg b on
(a.id = b.id and a.month = b.day and b.month = 1 and b.month = b.day and a.day = b.day)
left outer join alltypessmall c on
(c.id = a.id and a.month = c.id and a.month=c.month)
where a.id < 10;

when running test query_test.test_join_queries.TestJoinQueries.test_outer_joins

The reason is because a join condition contains =($0,$1), and both are on the
same side. It cannot be pushed because JOIN_CONDITION_PUSH does not allow
the push to the left side. So we need to skip over putting this into the
equality constraints.

Change-Id: Ie826a412bc5e650eb387cd4aa5aacf1a26535931